+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 3.4.4 (8/25/2015 - 8/26/2015)
 * rabbitmq: make listener use native thread
-* redis: track time to retrieve redis client from pool
+* redis: use internal pooling
 * rabbitmq: channel pooling (send message perf is 20x faster if not closing channel)
 
 ### 3.4.3 (8/19/2015 - 8/25/2015)